backend/pyproject.toml
[tool.poetry]
name = "reporting-backend"
version = "0.1.0"
description = "Reporting & Viz API"
authors = ["you <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
pydantic = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"
